타입, 상수 의미단위로 관리 (#32)
* chore: 절대경로 router 폴더 추가

* chore: 상수, 타입 의미단위로 관리
