+++ src/components/NavigationBar.tsx
@@ -3,8 +3,8 @@ import { ReactComponent as HomeIcon } from "@assets/icon/home.svg";
 import { ReactComponent as MessageIcon } from "@assets/icon/message.svg";
 import { ReactComponent as NewsIcon } from "@assets/icon/news.svg";
 import { ReactComponent as UserIcon } from "@assets/icon/user-circle.svg";
+import { PATH } from "@router/constant";
 import { HEIGHT } from "@styles/constants";
-import { PATH } from "constants/path";
 import { NavLink } from "react-router-dom";
 import styled from "styled-components";
