+++ src/components/common/Menu/type.ts
@@ -1,0 +1,7 @@
+export type MenuItemInfo = {
+  value: string;
+  onClick: (id?: number) => void;
+  itemId?: number;
+  isSelected?: boolean;
+  isWarning?: boolean;
+};
