+++ src/customTypes/index.ts
@@ -8,11 +8,3 @@ export type CategoryInfo = {
   name: string;
   imageUrl: string;
 };
-
-export type MenuItemInfo = {
-  value: string;
-  onClick: (id?: number) => void;
-  itemId?: number;
-  isSelected?: boolean;
-  isWarning?: boolean;
-};
